Add unit test for no self-dependencies
The output from the acceptance test identified one issue only: that the
current node was being included in the list of dependencies. I prefer to
keep the existing behaviour of including only additional dependencies,
as it is likely that a component will depend upon itself regardless of
whether it has any additional dependencies or whether that dependency is
made explicit through a circular series of relationships.